Cast to qreal for compatibility
svn path=/trunk/koffice/; revision=1096028

+++ plugins/textshape/Layout.cpp
@@ -1720,8 +1720,8 @@ void Layout::drawListItem(QPainter *painter, const QTextBlock &block, KoImageCol
         } else if (listStyle == KoListStyle::ImageItem && imageCollection) {
             QFontMetricsF fm(cf.font(), m_parent->paintDevice());
             qreal x = qMax(qreal(1), data->counterPosition().x());
-            qreal width = qMax(listFormat.doubleProperty(KoListStyle::Width), 1.0);
-            qreal height = qMax(listFormat.doubleProperty(KoListStyle::Height), 1.0);
+            qreal width = qMax(listFormat.doubleProperty(KoListStyle::Width), (qreal)1.0);
+            qreal height = qMax(listFormat.doubleProperty(KoListStyle::Height), (qreal)1.0);
             qreal y = data->counterPosition().y() + fm.ascent() - fm.xHeight()/2 - height/2; // centered
             qint64 key = listFormat.property(KoListStyle::BulletImageKey).value<qint64>();
             KoImageData idata;
